+++ internal/payment/domain/pricing_zones.go
@@ -6,16 +6,16 @@ import (
 
 // PricingZone represents a pricing zone for dynamic pricing
 type PricingZone struct {
-	ID                        string    `json:"id" db:"id"`
-	Country                   string    `json:"country" db:"country"`
-	ISOCode                   string    `json:"iso_code" db:"iso_code"`
-	Zone                      string    `json:"zone" db:"zone"`
-	ZoneName                  string    `json:"zone_name" db:"zone_name"`
-	WorldBankClassification   string    `json:"world_bank_classification" db:"world_bank_classification"`
-	GNIPerCapitaThreshold     string    `json:"gni_per_capita_threshold" db:"gni_per_capita_threshold"`
-	PricingMultiplier         float64   `json:"pricing_multiplier" db:"pricing_multiplier"`
-	CreatedAt                 time.Time `json:"created_at" db:"created_at"`
-	UpdatedAt                 time.Time `json:"updated_at" db:"updated_at"`
+	ID                      string    `json:"id" db:"id"`
+	Country                 string    `json:"country" db:"country"`
+	ISOCode                 string    `json:"iso_code" db:"iso_code"`
+	Zone                    string    `json:"zone" db:"zone"`
+	ZoneName                string    `json:"zone_name" db:"zone_name"`
+	WorldBankClassification string    `json:"world_bank_classification" db:"world_bank_classification"`
+	GNIPerCapitaThreshold   string    `json:"gni_per_capita_threshold" db:"gni_per_capita_threshold"`
+	PricingMultiplier       float64   `json:"pricing_multiplier" db:"pricing_multiplier"`
+	CreatedAt               time.Time `json:"created_at" db:"created_at"`
+	UpdatedAt               time.Time `json:"updated_at" db:"updated_at"`
 }
 
 // PricingZoneRequest represents a request for pricing zone operations
@@ -23,14 +23,14 @@ type PricingZoneRequest struct {
 	Country   string  `json:"country,omitempty"`
 	ISOCode   string  `json:"iso_code,omitempty"`
 	Zone      string  `json:"zone,omitempty"`
-	BasePrice int64   `json:"base_price"` // Base price in cents
+	BasePrice float64 `json:"base_price"` // Base price in dollars
 }
 
 // PricingZoneResponse represents the response from pricing zone operations
 type PricingZoneResponse struct {
 	Zone              *PricingZone `json:"zone,omitempty"`
-	BasePrice         int64        `json:"base_price"`         // Original base price in cents
-	AdjustedPrice     int64        `json:"adjusted_price"`     // Price after multiplier in cents
+	BasePrice         float64      `json:"base_price"`         // Original base price in dollars
+	AdjustedPrice     float64      `json:"adjusted_price"`     // Price after multiplier in dollars
 	PricingMultiplier float64      `json:"pricing_multiplier"` // Applied multiplier
 	Currency          string       `json:"currency"`           // Currency code
 }
@@ -39,9 +39,9 @@ type PricingZoneResponse struct {
 type ZoneType string
 
 const (
-	ZoneTypePremium  ZoneType = "A" // Premium - 100% of base price
-	ZoneTypeMidHigh  ZoneType = "B" // Mid-High - 70% of base price
-	ZoneTypeMidLow   ZoneType = "C" // Mid-Low - 40% of base price
+	ZoneTypePremium   ZoneType = "A" // Premium - 100% of base price
+	ZoneTypeMidHigh   ZoneType = "B" // Mid-High - 70% of base price
+	ZoneTypeMidLow    ZoneType = "C" // Mid-Low - 40% of base price
 	ZoneTypeLowIncome ZoneType = "D" // Low-Income - 20% of base price
 )
 
@@ -78,13 +78,13 @@ func GetDefaultMultiplier(zoneType ZoneType) float64 {
 }
 
 // CalculateAdjustedPrice calculates the adjusted price based on the pricing multiplier
-func (pz *PricingZone) CalculateAdjustedPrice(basePrice int64) int64 {
+func (pz *PricingZone) CalculateAdjustedPrice(basePrice float64) float64 {
 	if pz.PricingMultiplier <= 0 {
 		return basePrice
 	}
-	
-	adjustedPrice := float64(basePrice) * pz.PricingMultiplier
-	return int64(adjustedPrice)
+
+	adjustedPrice := basePrice * pz.PricingMultiplier
+	return adjustedPrice
 }
 
 // IsValidZone checks if the zone is valid
